Slightly better function

+++ color.go
@@ -23,9 +23,9 @@ Return:
 	Distance
 */
 func Distance(lhs, rhs color.RGBA) float64 {
-	r := float64(lhs.R) - float64(rhs.R)
-	g := float64(lhs.G) - float64(rhs.G)
-	b := float64(lhs.B) - float64(rhs.B)
+	r := float64(lhs.R - rhs.R)
+	g := float64(lhs.G - rhs.G)
+	b := float64(lhs.B - rhs.B)
 	l := Luminance(lhs) - Luminance(rhs)
 	return math.Sqrt(r*r + g*g + b*b + l*l)
 }
